requirements.txt
# AskTube - YouTube Learning Assistant
# Python Dependencies

# Core dependencies
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# LangChain and LLM
langchain>=0.1.0
langchain-google-genai>=1.0.0
langchain-core>=0.1.0
langchain-community>=0.0.20

# YouTube transcript extraction
youtube-transcript-api>=0.6.0

# Google Generative AI
google-generativeai>=0.3.0

# PDF generation (optional but recommended)
weasyprint>=60.0

# Optional: For better console output on Windows
colorama>=0.4.6

# RAG components
sentence-transformers>=2.2.2
faiss-cpu>=1.8.0.post1
huggingface_hub<1.0

# Ollama Python client for local LLM backend
ollama>=0.1.8

# FastAPI backend
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
motor>=3.3.2
pymongo[srv]>=4.8.0
python-multipart>=0.0.9
# Optional structured logging
loguru>=0.7.2
pytest>=8.3.0
httpx>=0.27.0
pytest-asyncio>=0.23.0
anyio>=4.4.0
